requirements.txt
django
djangorestframework
pandas
psycopg2
# psycopg2-binary